Make context management match underlying GLFW
The bind()/unbind() and its attached thread concept is not suitable with
the context management of GLFW.

Binding was meant to be equivalent with making context current with
glfwMakeContextCurrent() but with an addition of tracking the calling
thread inside Window. This however won't work correctly since when
changing current context the previous window should unbind() itself
before the new one bind() itself. In the case unbind() not called for
previous window, both will have the same attached thread which should
not happen!

In GLFW there is no such thing anyway so removing these interfaces is
best.